+++ TheMovieDB/Package.swift
@@ -28,7 +28,6 @@ let package = Package(
     dependencies: [
         .package(url: "https://github.com/evgenyneu/keychain-swift.git", from: "14.0.0"),
         .package(url: "https://github.com/realm/realm-swift.git", from: "10.0.0"),
-        .package(url: "https://github.com/pointfreeco/combine-schedulers.git", from: "0.9.1"),
         .package(url: "https://github.com/airbnb/lottie-spm.git", from: "4.1.2"),
     ],
     targets: [
@@ -91,7 +90,6 @@ let package = Package(
                     "Network",
                     "MovieListFeatureInterface",
                     "UI",
-                    .product(name: "CombineSchedulers", package: "combine-schedulers"),
                     .product(name: "Lottie", package: "lottie-spm")
                 ]
             ),
